pyproject.toml
[tool.poetry]
authors = ["Daniel Georg", "Julian Obst", "Vladimir Yussupov"]
description = "Qunicorn Core for SeQuenC"
license = "Apache-2.0"
name = "qunicorn_core"
readme = "README.md"
version = "0.2.0"

[tool.poetry.dependencies]
Flask-Cors = "^3.0.10"
Flask-JWT-Extended = "^4.4.4"
Flask-Migrate = "^4.0.4"
Flask-SQLAlchemy = "^3.0.3"
SQLAlchemy = "^2.0.4"
flask = { extras = ["dotenv"], version = "^2.2.3" }
flask-babel = "^3.0.0"
flask-smorest = "^0.40.0"
python = "~3.11.0"
tomli = "^2.0.0"
celery = "^5.2.7"
redis = "^4.5.4"
qiskit = "^0.44.0"
qiskit-ibm-provider = "^0.6.0"
qiskit-ibm-runtime = "^0.11.1"
PyYAML = "^6.0"
mocker = "^1.1.1"
amazon-braket-sdk = "^1.49.1"
qrisp = "^0.2.1"

[tool.poetry.dev-dependencies]
Sphinx = "^5.0.0"
black = "^23.1.0"
flake8 = "^6.0.0"
flake8-bugbear = "^23.2.13"
flake8-docstrings = "^1.7.0"
invoke = "^2.0.0"
myst-parser = "^0.18.1"
pip-licenses = "^4.0.3"
pytest = "^7.2.0"
pytest-mock = "^3.11.1"
sphinx-click = "^4.4.0"
sphinxcontrib-redoc = "^1.6.0"

[tool.poetry.scripts]
#flask = 'flask.cli:main'

[tool.black]
include = '\.pyi?$'
line-length = 120
target-version = ["py38"]

[tool.isort]
atomic = true
group_by_package = true
honor_noqa = true
multi_line_output = 3
profile = "black"
reverse_relative = true

[tool.sphinx]
copyright-year = 2023
enable-autodoc = true
enable-autosectionlabel = true
# documentation for myst: <https://myst-parser.readthedocs.io/en/latest>
enable-markdown = "myst"
enable-todo = true
# update gitignore if you want to edit docs/changelog.md manually!
include-changelog = true
# update gitignore if you want to edit docs/readme.md manually!
include-readme = true

[tool.sphinx.intersphinx-mapping]
python = ["https://docs.python.org/3"]

[tool.sphinx.myst]
# settings documentation: <https://myst-parser.readthedocs.io/en/latest/syntax/optional.html>
# create anchors for h1 and h2
extensions = [
    #"amsmath",
    "colon_fence",
    "deflist",
    "dollarmath",
    "html_admonition",
    "html_image", #"linkify", # requires installation of linkify-it-py
    "replacements",
    "smartquotes",
    "substitution",
    "tasklist",
]
heading_anchors = 2
# substitutions = {"key": "substitution value"}

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
